Filter task separated in its own task dptdpt-filter
It feeds
- match-reco-gen independent separated from the filter-qa task
- dptdptcorrelations independent
- dptdpt-filter-qa independent separated from the match-reco-gen task
