+++ PWGCF/DataModel/DptDptFiltered.h
@@ -1,0 +1,68 @@
+// Copyright 2019-2020 CERN and copyright holders of ALICE O2.
+// See https://alice-o2.web.cern.ch/copyright for details of the copyright holders.
+// All rights not expressly granted are reserved.
+//
+// This software is distributed under the terms of the GNU General Public
+// License v3 (GPL Version 3), copied verbatim in the file "COPYING".
+//
+// In applying this license CERN does not waive the privileges and immunities
+// granted to it by virtue of its status as an Intergovernmental Organization
+// or submit itself to any jurisdiction.
+#ifndef O2_ANALYSIS_DPTDPTFILTERED_H
+#define O2_ANALYSIS_DPTDPTFILTERED_H
+
+#include "Framework/ASoA.h"
+#include "Framework/AnalysisDataModel.h"
+
+namespace o2
+{
+namespace aod
+{
+/* we have to change from int to bool when bool columns work properly */
+namespace dptdptfilter
+{
+DECLARE_SOA_COLUMN(EventAccepted, eventaccepted, uint8_t); //! If the collision/event has been accepted or not
+DECLARE_SOA_COLUMN(EventCentMult, centmult, float);        //! The centrality/multiplicity pecentile
+} // namespace dptdptfilter
+DECLARE_SOA_TABLE(AcceptedEvents, "AOD", "ACCEPTEDEVENTS", //! Accepted reconstructed collisions/events filtered table
+                  o2::soa::Index<>,
+                  collision::BCId,
+                  collision::PosZ,
+                  dptdptfilter::EventAccepted,
+                  dptdptfilter::EventCentMult);
+using AcceptedEvent = AcceptedEvents::iterator;
+DECLARE_SOA_TABLE(AcceptedTrueEvents, "AOD", "ACCTRUEEVENTS", //! Accepted generated collisions/events filtered table
+                  o2::soa::Index<>,
+                  collision::BCId,
+                  mccollision::PosZ,
+                  dptdptfilter::EventAccepted,
+                  dptdptfilter::EventCentMult);
+using AcceptedTrueEvent = AcceptedTrueEvents::iterator;
+namespace dptdptfilter
+{
+DECLARE_SOA_INDEX_COLUMN(AcceptedEvent, event);                      //! Reconstructed collision/event
+DECLARE_SOA_INDEX_COLUMN(AcceptedTrueEvent, mcevent);                //! Generated collision/event
+DECLARE_SOA_COLUMN(TrackacceptedAsOne, trackacceptedasone, uint8_t); //! Track accepted as type one
+DECLARE_SOA_COLUMN(TrackacceptedAsTwo, trackacceptedastwo, uint8_t); //! Track accepted as type two
+DECLARE_SOA_COLUMN(Pt, pt, float);                                   //! The track transverse momentum
+DECLARE_SOA_COLUMN(Eta, eta, float);                                 //! The track pseudorapidity
+DECLARE_SOA_COLUMN(Phi, phi, float);                                 //! The track azimuthal angle
+} // namespace dptdptfilter
+DECLARE_SOA_TABLE(ScannedTracks, "AOD", "SCANNEDTRACKS", //! The reconstructed tracks filtered table
+                  dptdptfilter::AcceptedEventId,
+                  dptdptfilter::TrackacceptedAsOne,
+                  dptdptfilter::TrackacceptedAsTwo,
+                  dptdptfilter::Pt,
+                  dptdptfilter::Eta,
+                  dptdptfilter::Phi);
+DECLARE_SOA_TABLE(ScannedTrueTracks, "AOD", "SCANTRUETRACKS", //! The generated particles filtered table
+                  dptdptfilter::AcceptedTrueEventId,
+                  dptdptfilter::TrackacceptedAsOne,
+                  dptdptfilter::TrackacceptedAsTwo,
+                  dptdptfilter::Pt,
+                  dptdptfilter::Eta,
+                  dptdptfilter::Phi);
+} // namespace aod
+} // namespace o2
+
+#endif // O2_ANALYSIS_DPTDPTFILTERED_H
