+++ PWGCF/Tasks/CMakeLists.txt
@@ -14,6 +14,11 @@ o2physics_add_dpl_workflow(dptdptcorrelations
                     PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::PWGCFCore
                     COMPONENT_NAME Analysis)
 
+o2physics_add_dpl_workflow(dptdpt-filter-qa
+                    SOURCES dptdptfilterqa.cxx
+                    PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::PWGCFCore
+                    COMPONENT_NAME Analysis)
+
 o2physics_add_dpl_workflow(match-reco-gen
                     SOURCES match-reco-gen.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::PWGCFCore
